refactor(frontend): move to separate file
