+++ apps/frontend/app/components/common.tsx
@@ -20,6 +20,7 @@ import {
 	Modal,
 	MultiSelect,
 	Paper,
+	RingProgress,
 	SimpleGrid,
 	Skeleton,
 	Stack,
@@ -28,6 +29,7 @@ import {
 	Title,
 	Tooltip,
 	rem,
+	useMantineTheme,
 } from "@mantine/core";
 import { useDebouncedValue, useDidUpdate, useDisclosure } from "@mantine/hooks";
 import {
@@ -40,29 +42,36 @@ import {
 import {
 	EntityLot,
 	GridPacking,
-	type MediaLot,
+	MediaLot,
 	type MediaSource,
 	type ReviewItem,
+	type UserAnalytics,
 	UserReviewScale,
 } from "@ryot/generated/graphql/backend/graphql";
 import {
 	changeCase,
+	formatQuantityWithCompactNotation,
 	getInitials,
+	humanizeDuration,
 	isNumber,
 	isString,
 	snakeCase,
 } from "@ryot/ts-utils";
 import {
 	IconArrowBigUp,
+	IconBarbell,
 	IconCheck,
 	IconEdit,
 	IconExternalLink,
 	IconFilterOff,
+	IconFriends,
 	IconMoodEmpty,
 	IconMoodHappy,
 	IconMoodSad,
 	IconRefresh,
+	IconScaleOutline,
 	IconSearch,
+	IconServer,
 	IconStarFilled,
 	IconTrash,
 	IconX,
@@ -71,16 +80,18 @@ import { useQuery } from "@tanstack/react-query";
 import clsx from "clsx";
 import Cookies from "js-cookie";
 import type { ReactNode, Ref } from "react";
-import { useState } from "react";
+import { Fragment, useState } from "react";
 import { $path } from "remix-routes";
 import type { DeepPartial } from "ts-essentials";
 import { match } from "ts-pattern";
 import { withQuery } from "ufo";
 import {
+	MediaColors,
 	PRO_REQUIRED_MESSAGE,
 	ThreePointSmileyRating,
 	convertDecimalToThreePointSmiley,
 	dayjsLib,
+	getMetadataIcon,
 	getSurroundingElements,
 	openConfirmationModal,
 	reviewYellow,
@@ -90,10 +101,12 @@ import {
 	useConfirmSubmit,
 	useCoreDetails,
 	useFallbackImageUrl,
+	useGetMantineColors,
 	useGetRandomMantineColor,
 	useUserCollections,
 	useUserDetails,
 	useUserPreferences,
+	useUserUnitSystem,
 } from "~/lib/hooks";
 import { useReviewEntity } from "~/lib/state/media";
 import type { action } from "~/routes/actions";
@@ -102,6 +115,7 @@ import {
 	ExerciseDisplayItem,
 	WorkoutDisplayItem,
 	WorkoutTemplateDisplayItem,
+	displayWeightWithUnit,
 } from "./fitness";
 import {
 	MetadataDisplayItem,
@@ -941,5 +955,364 @@ export const DisplayCollection = (props: {
 				</Flex>
 			</Form>
 		</Badge>
+	);
+};
+
+export const DisplaySummarySection = ({
+	latestUserSummary,
+}: {
+	latestUserSummary: UserAnalytics["activities"]["items"][0];
+}) => {
+	const userPreferences = useUserPreferences();
+	const unitSystem = useUserUnitSystem();
+	const theme = useMantineTheme();
+
+	return (
+		<SimpleGrid
+			cols={{ base: 1, sm: 2, md: 3 }}
+			style={{ alignItems: "center" }}
+			spacing="xs"
+		>
+			<DisplayStatForMediaType
+				lot={MediaLot.Movie}
+				data={[
+					{
+						label: "Movies",
+						value: latestUserSummary.movieCount,
+						type: "number",
+					},
+					{
+						label: "Runtime",
+						value: latestUserSummary.totalMovieDuration,
+						type: "duration",
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.Music}
+				data={[
+					{
+						label: "Songs",
+						value: latestUserSummary.musicCount,
+						type: "number",
+					},
+					{
+						label: "Runtime",
+						value: latestUserSummary.totalMusicDuration,
+						type: "duration",
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.Show}
+				data={[
+					{
+						label: "Show episodes",
+						value: latestUserSummary.showCount,
+						type: "number",
+					},
+					{
+						label: "Runtime",
+						value: latestUserSummary.totalShowDuration,
+						type: "duration",
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.VideoGame}
+				data={[
+					{
+						label: "Video games",
+						value: latestUserSummary.videoGameCount,
+						type: "number",
+					},
+					{
+						label: "Runtime",
+						value: latestUserSummary.totalVideoGameDuration,
+						type: "duration",
+						hideIfZero: true,
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.VisualNovel}
+				data={[
+					{
+						label: "Visual Novels",
+						value: latestUserSummary.visualNovelCount,
+						type: "number",
+					},
+					{
+						label: "Runtime",
+						value: latestUserSummary.totalVisualNovelDuration,
+						type: "duration",
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.AudioBook}
+				data={[
+					{
+						label: "Audio books",
+						value: latestUserSummary.audioBookCount,
+						type: "number",
+					},
+					{
+						label: "Runtime",
+						value: latestUserSummary.totalAudioBookDuration,
+						type: "duration",
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.Book}
+				data={[
+					{
+						label: "Books",
+						value: latestUserSummary.bookCount,
+						type: "number",
+					},
+					{
+						label: "Pages",
+						value: latestUserSummary.totalBookPages,
+						type: "number",
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.Podcast}
+				data={[
+					{
+						label: "Podcasts",
+						value: latestUserSummary.podcastCount,
+						type: "number",
+					},
+					{
+						label: "Runtime",
+						value: latestUserSummary.totalPodcastDuration,
+						type: "duration",
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.Manga}
+				data={[
+					{
+						label: "Manga",
+						value: latestUserSummary.mangaCount,
+						type: "number",
+					},
+				]}
+			/>
+			<DisplayStatForMediaType
+				lot={MediaLot.Anime}
+				data={[
+					{
+						label: "Anime",
+						value: latestUserSummary.animeCount,
+						type: "number",
+					},
+				]}
+			/>
+			{userPreferences.featuresEnabled.media.enabled ? (
+				<>
+					<ActualDisplayStat
+						icon={<IconServer />}
+						lot="Metadata stats"
+						color={theme.colors.grape[8]}
+						data={[
+							{
+								label: "Media",
+								value: latestUserSummary.totalMetadataCount,
+								type: "number",
+							},
+							{
+								label: "Reviews",
+								value: latestUserSummary.totalMetadataReviewCount,
+								type: "number",
+								hideIfZero: true,
+							},
+						]}
+					/>
+					{userPreferences.featuresEnabled.media.people ? (
+						<UnstyledLink
+							to={$path("/media/people/:action", {
+								action: "list",
+							})}
+						>
+							<ActualDisplayStat
+								icon={<IconFriends />}
+								lot="People stats"
+								color={theme.colors.red[9]}
+								data={[
+									{
+										label: "People Reviewed",
+										value: latestUserSummary.totalPersonReviewCount,
+										type: "number",
+										hideIfZero: true,
+									},
+								]}
+							/>
+						</UnstyledLink>
+					) : null}
+				</>
+			) : null}
+			{userPreferences.featuresEnabled.fitness.enabled ? (
+				<UnstyledLink
+					to={$path("/fitness/:entity/list", {
+						entity: "workouts",
+					})}
+				>
+					<ActualDisplayStat
+						icon={<IconBarbell stroke={1.3} />}
+						lot="Workouts"
+						color={theme.colors.teal[2]}
+						data={[
+							{
+								label: "Workouts",
+								value: latestUserSummary.workoutCount,
+								type: "number",
+							},
+							{
+								label: "Runtime",
+								value: latestUserSummary.totalWorkoutDuration,
+								type: "duration",
+							},
+							{
+								label: "Runtime",
+								value: displayWeightWithUnit(
+									unitSystem,
+									latestUserSummary.totalWorkoutWeight,
+									true,
+								),
+								type: "string",
+							},
+						]}
+					/>
+				</UnstyledLink>
+			) : null}
+			{userPreferences.featuresEnabled.fitness.enabled ? (
+				<ActualDisplayStat
+					icon={<IconScaleOutline stroke={1.3} />}
+					lot="Fitness"
+					color={theme.colors.yellow[5]}
+					data={[
+						{
+							label: "Measurements",
+							value: latestUserSummary.userMeasurementCount,
+							type: "number",
+							hideIfZero: true,
+						},
+					]}
+				/>
+			) : null}
+		</SimpleGrid>
+	);
+};
+
+const ActualDisplayStat = (props: {
+	icon: ReactNode;
+	lot: string;
+	data: Array<{
+		type: "duration" | "number" | "string";
+		label: string;
+		value: number | string;
+		hideIfZero?: true;
+	}>;
+	color?: string;
+}) => {
+	const colors = useGetMantineColors();
+
+	return (
+		<Flex align="center">
+			<RingProgress
+				size={60}
+				thickness={4}
+				sections={[]}
+				label={<Center>{props.icon}</Center>}
+				rootColor={props.color ?? colors[11]}
+			/>
+			<Flex wrap="wrap" ml="xs">
+				{props.data.map((d, idx) => (
+					<Fragment key={idx.toString()}>
+						{isNumber(d.type) && d.value === 0 && d.hideIfZero ? undefined : (
+							<Box mx="xs" data-stat-stringified={JSON.stringify(d)}>
+								<Text
+									fw={d.label !== "Runtime" ? "bold" : undefined}
+									display="inline"
+									fz={{ base: "md", md: "sm", xl: "md" }}
+								>
+									{match(d.type)
+										.with("string", () => d.value)
+										.with("duration", () =>
+											humanizeDuration(
+												dayjsLib
+													.duration(Number(d.value), "minutes")
+													.asMilliseconds(),
+												{
+													round: true,
+													largest: 3,
+												},
+											),
+										)
+										.with("number", () =>
+											formatQuantityWithCompactNotation(Number(d.value)),
+										)
+										.exhaustive()}
+								</Text>
+								<Text
+									display="inline"
+									ml="4px"
+									fz={{ base: "md", md: "sm", xl: "md" }}
+								>
+									{d.label === "Runtime" ? "" : d.label}
+								</Text>
+							</Box>
+						)}
+					</Fragment>
+				))}
+			</Flex>
+		</Flex>
+	);
+};
+
+const DisplayStatForMediaType = (props: {
+	lot: MediaLot;
+	data: Array<{
+		type: "duration" | "number";
+		label: string;
+		value: number;
+		hideIfZero?: true;
+	}>;
+}) => {
+	const userPreferences = useUserPreferences();
+	const isEnabled = userPreferences.featuresEnabled.media.specific.includes(
+		props.lot,
+	);
+	const Icon = getMetadataIcon(props.lot);
+	const icon = <Icon size={24} stroke={1.5} />;
+
+	return isEnabled && userPreferences.featuresEnabled.media.enabled ? (
+		<UnstyledLink
+			to={$path("/media/:action/:lot", {
+				action: "list",
+				lot: props.lot.toLowerCase(),
+			})}
+		>
+			<ActualDisplayStat
+				data={props.data}
+				icon={icon}
+				lot={props.lot.toString()}
+				color={MediaColors[props.lot]}
+			/>
+		</UnstyledLink>
+	) : null;
+};
+
+const UnstyledLink = (props: { children: ReactNode; to: string }) => {
+	return (
+		<Link to={props.to} style={{ all: "unset", cursor: "pointer" }}>
+			{props.children}
+		</Link>
 	);
 };
